refactor: Split CharacterModel into three parts
They are:
* Character model itself, dealing with "class" things
* CharacterStats, dealing with stuff that change dynamically like health, mana (and I'd like to fold CombatXp into this as well)
* StatusEffect, dealing with, eh, status effects

This is in preparation for cleaning up CharacterModel (the "class" stuff)
