+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -10,14 +10,15 @@ import com.wynntils.models.abilities.ShamanMaskModel;
 import com.wynntils.models.abilities.ShamanTotemModel;
 import com.wynntils.models.character.CharacterModel;
 import com.wynntils.models.character.CharacterSelectionModel;
-import com.wynntils.models.character.PlayerInventoryModel;
+import com.wynntils.models.characterstats.CharacterStatsModel;
+import com.wynntils.models.characterstats.CombatXpModel;
 import com.wynntils.models.containers.ContainerModel;
 import com.wynntils.models.containers.LootChestModel;
+import com.wynntils.models.containers.PlayerInventoryModel;
 import com.wynntils.models.damage.DamageModel;
 import com.wynntils.models.discoveries.DiscoveryModel;
 import com.wynntils.models.elements.ElementModel;
 import com.wynntils.models.emeralds.EmeraldModel;
-import com.wynntils.models.experience.CombatXpModel;
 import com.wynntils.models.favorites.FavoritesModel;
 import com.wynntils.models.gear.GearModel;
 import com.wynntils.models.gear.GearTooltipModel;
@@ -38,6 +39,7 @@ import com.wynntils.models.quests.QuestModel;
 import com.wynntils.models.rewards.RewardsModel;
 import com.wynntils.models.spells.SpellModel;
 import com.wynntils.models.stats.StatModel;
+import com.wynntils.models.statuseffects.StatusEffectModel;
 import com.wynntils.models.territories.GuildAttackTimerModel;
 import com.wynntils.models.territories.TerritoryModel;
 import com.wynntils.models.territories.TowerAuraTimerModel;
@@ -51,6 +53,7 @@ public final class Models {
     public static final TowerAuraTimerModel TowerAuraTimer = new TowerAuraTimerModel();
     public static final BombBellModel BombBell = new BombBellModel();
     public static final BossBarModel BossBar = new BossBarModel();
+    public static final CharacterModel Character = new CharacterModel();
     public static final CharacterSelectionModel CharacterSelection = new CharacterSelectionModel();
     public static final CompassModel Compass = new CompassModel();
     public static final ContainerModel Container = new ContainerModel();
@@ -64,13 +67,14 @@ public final class Models {
     public static final RewardsModel Rewards = new RewardsModel();
     public static final ServerListModel ServerList = new ServerListModel();
     public static final StatModel Stat = new StatModel();
+    public static final StatusEffectModel StatusEffect = new StatusEffectModel();
     public static final TerritoryModel Territory = new TerritoryModel();
     public static final WorldStateModel WorldState = new WorldStateModel();
     public static final WynnItemModel WynnItem = new WynnItemModel();
 
     // Models with dependencies, ordered alphabetically as far as possible
     public static final CombatXpModel CombatXp = new CombatXpModel(WorldState);
-    public static final CharacterModel Character = new CharacterModel(CombatXp);
+    public static final CharacterStatsModel CharacterStats = new CharacterStatsModel(CombatXp);
     public static final FriendsModel Friends = new FriendsModel(WorldState);
     public static final GearModel Gear = new GearModel(Element, Stat, WynnItem);
     public static final HadesModel Hades = new HadesModel(Character, WorldState);
