+++ common/src/main/java/com/wynntils/features/combat/LowHealthVignetteFeature.java
@@ -47,7 +47,7 @@ public class LowHealthVignetteFeature extends UserFeature {
 
     @SubscribeEvent
     public void onTick(TickEvent event) {
-        float healthProgress = (float) Models.Character.getHealth().getProgress();
+        float healthProgress = (float) Models.CharacterStats.getHealth().getProgress();
         float threshold = lowHealthPercentage / 100f;
         shouldRender = false;
