+++ common/src/main/java/com/wynntils/models/characterstats/CombatXpModel.java
@@ -2,13 +2,13 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.experience;
+package com.wynntils.models.characterstats;
 
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Model;
 import com.wynntils.core.components.Models;
 import com.wynntils.mc.event.SetXpEvent;
-import com.wynntils.models.experience.event.CombatXpGainEvent;
+import com.wynntils.models.characterstats.event.CombatXpGainEvent;
 import com.wynntils.models.worlds.WorldStateModel;
 import com.wynntils.models.worlds.event.WorldStateEvent;
 import com.wynntils.models.worlds.type.WorldState;
