+++ common/src/main/java/com/wynntils/models/characterstats/actionbar/SprintSegment.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.character.actionbar;
+package com.wynntils.models.characterstats.actionbar;
 
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.handlers.actionbar.ActionBarSegment;
