+++ common/src/main/java/com/wynntils/models/characterstats/event/CombatXpGainEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.experience.event;
+package com.wynntils.models.characterstats.event;
 
 import net.minecraftforge.eventbus.api.Event;
