+++ common/src/main/java/com/wynntils/models/players/hades/HadesModel.java
@@ -202,7 +202,7 @@ public final class HadesModel extends Model {
             float pZ = (float) player.getZ();
 
             PlayerStatus newStatus =
-                    new PlayerStatus(pX, pY, pZ, Models.Character.getHealth(), Models.Character.getMana());
+                    new PlayerStatus(pX, pY, pZ, Models.CharacterStats.getHealth(), Models.CharacterStats.getMana());
 
             if (newStatus.equals(lastSentStatus)) {
                 tickCountUntilUpdate = 1;
