+++ common/src/main/java/com/wynntils/models/statuseffects/type/StatusEffect.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.character.type;
+package com.wynntils.models.statuseffects.type;
 
 public class StatusEffect {
     private final String fullName;
